+++ integrations/javascript/examples/webview2-renderer.cjs
@@ -1,6 +1,6 @@
 'use strict';
 
-const { createWebView2RendererClient } = require('../src/webview2.cjs');
+const { createWebView2RendererClient } = require('..');
 
 function createPatrisClient(webview = window.chrome.webview) {
   return createWebView2RendererClient({ webview });